+++ CompareLengths.java
@@ -4,7 +4,8 @@ class CompareLengths implements Comparator<Caves>{
 	
 	public int compare(Caves a, Caves b)
 	{
-		return Double.compare(a.length, b.length);
+		return Long.compare(a.length, b.length);
+		//return Double.compare(a.length, b.length);
 	}
 
 }
